mode_RPiHQ_mean.hpp: fix return type

+++ src/include/mode_RPiHQ_mean.h
@@ -30,6 +30,6 @@ struct modeMeanSetting {
 };
 
 
-int RPiHQcalcMean(const char* fileName, int asiExposure, double asiGain, raspistillSetting &currentRaspistillSetting, modeMeanSetting &currentModeMeanSetting);
+float RPiHQcalcMean(const char* fileName, int asiExposure, double asiGain, raspistillSetting &currentRaspistillSetting, modeMeanSetting &currentModeMeanSetting);
 
 #endif
